fix: store session lifetime inside session file
