+++ cache.go
@@ -1,9 +1,6 @@
 package zabbix
 
 import (
-	"encoding/json"
-	"io/ioutil"
-	"os"
 	"time"
 )
 
@@ -23,96 +20,4 @@ type SessionAbstractCache interface {
 
 	// Flush removes cached session
 	Flush() error
-}
-
-// SessionFileCache is Zabbix session filesystem cache.
-type SessionFileCache struct {
-	filePath        string
-	sessionLifeTime time.Duration
-	filePermissions uint32
-}
-
-// SetFilePath sets Zabbix session cache file path. Default value is "./zabbix_session"
-func (c *SessionFileCache) SetFilePath(filePath string) *SessionFileCache {
-	c.filePath = filePath
-	return c
-}
-
-// SetFilePermissions sets permissions for a session file. Default value is 0655.
-func (c *SessionFileCache) SetFilePermissions(permissions uint32) *SessionFileCache {
-	c.filePermissions = permissions
-	return c
-}
-
-// SetSessionLifetime sets lifetime in seconds of cached Zabbix session. Default value is 4 hours.
-func (c *SessionFileCache) SetSessionLifetime(d time.Duration) {
-	c.sessionLifeTime = d
-}
-
-// SaveSession saves session to a cache
-func (c *SessionFileCache) SaveSession(session *Session) error {
-	serialized, err := json.Marshal(session)
-
-	if err != nil {
-		return err
-	}
-
-	return ioutil.WriteFile(c.filePath, []byte(serialized), os.FileMode(c.filePermissions))
-}
-
-// GetSession returns cached Zabbix session
-func (c *SessionFileCache) GetSession() (*Session, error) {
-	contents, err := ioutil.ReadFile(c.filePath)
-
-	if err != nil {
-		return nil, err
-	}
-
-	var session Session
-
-	if err := json.Unmarshal(contents, &session); err != nil {
-		return nil, err
-	}
-
-	return &session, err
-}
-
-// HasSession checks if any valid Zabbix session has been cached and available
-func (c *SessionFileCache) HasSession() bool {
-	stats, err := os.Stat(c.filePath)
-
-	// Try to check if session file exists and modify date
-	if os.IsNotExist(err) {
-		return false
-	}
-
-	// If file exists, check mod time as used in original portmon-sync
-	now := time.Now().Unix()
-	lastMod := stats.ModTime().Unix()
-	timeDiff := now - lastMod
-
-	// Check session TTL by time diff
-	if timeDiff > int64(c.sessionLifeTime) {
-
-		// Delete outdated session file
-		os.Remove(c.filePath)
-
-		return false
-	}
-
-	return true
-}
-
-// Flush removes a cached session
-func (c *SessionFileCache) Flush() error {
-	return os.Remove(c.filePath)
-}
-
-// NewSessionFileCache creates a new instance of session file system cache
-func NewSessionFileCache() *SessionFileCache {
-	return &SessionFileCache{
-		filePath:        "./zabbix_session",
-		sessionLifeTime: 14400, // Default TTL is 4 hours
-		filePermissions: 0600,
-	}
 }
